+++ src/connection/client.spec.ts
@@ -191,6 +191,44 @@ describe('DatabaseClient', () => {
     })
   })
 
+  describe('.insert()', () => {
+    const song = { name: 'Take On Me', artist: 'A-ha', rating: 5 }
+
+    it('inserts the given record', async () => {
+      const newSong = { id: 1, ...song }
+
+      const { client } = mkClient()
+      jest.spyOn(client, 'query').mockResolvedValue([newSong])
+
+      await expect(client.insert('song', song)).resolves.toBe(newSong)
+
+      expect(client.query).toHaveBeenCalledWith(
+        expect.sqlMatching({
+          text: `
+            INSERT INTO "song" ("name","artist","rating")
+            VALUES ($1,$2,$3)
+            RETURNING *
+          `,
+          values: ['Take On Me', 'A-ha', 5],
+        }),
+      )
+    })
+
+    it('returns null if no rows come back', async () => {
+      const { client } = mkClient()
+      jest.spyOn(client, 'query').mockResolvedValue([])
+
+      await expect(client.insert('song', song)).resolves.toBeNull()
+    })
+
+    it('errors if multiple rows come back', async () => {
+      const { client } = mkClient()
+      jest.spyOn(client, 'query').mockResolvedValue([{}, {}])
+
+      await expect(client.insert('song', song)).rejects.toThrow()
+    })
+  })
+
   describe('.insertAll()', () => {
     it('inserts the given records', async () => {
       const songs = [
@@ -207,6 +245,7 @@ describe('DatabaseClient', () => {
           text: `
             INSERT INTO "song" ("name","artist","rating")
             VALUES ($1,$2,$3)
+            RETURNING *
           `,
           values: ['Take On Me', 'A-ha', 5],
         }),
@@ -214,6 +253,7 @@ describe('DatabaseClient', () => {
           text: `
             INSERT INTO "song" ("name","artist")
             VALUES ($1,$2)
+            RETURNING *
           `,
           values: ['Separate Ways', 'Journey'],
         }),
@@ -241,19 +281,19 @@ describe('DatabaseClient', () => {
       expect(mockQuery).toHaveBeenCalledWith(expect.sqlMatching('BEGIN'))
       expect(mockQuery).toHaveBeenCalledWith(
         expect.sqlMatching({
-          text: `INSERT INTO "song" ("name") VALUES ($1)`,
+          text: `INSERT INTO "song" ("name") VALUES ($1) RETURNING *`,
           values: ['Take On Me'],
         }),
       )
       expect(mockQuery).toHaveBeenCalledWith(
         expect.sqlMatching({
-          text: `INSERT INTO "song" ("name") VALUES ($1)`,
+          text: `INSERT INTO "song" ("name") VALUES ($1) RETURNING *`,
           values: ['UNKNOWN'],
         }),
       )
       expect(mockQuery).not.toHaveBeenCalledWith(
         expect.sqlMatching({
-          text: `INSERT INTO "song" ("name") VALUES ($1)`,
+          text: `INSERT INTO "song" ("name") VALUES ($1) RETURNING *`,
           values: ['Separate Ways'],
         }),
       )